Make completeChallenge perform bank transaction, and cleanup code

+++ src/main/java/idatt2106/systemutvikling/sparesti/service/ChallengeService.java
@@ -2,13 +2,18 @@ package idatt2106.systemutvikling.sparesti.service;
 
 import idatt2106.systemutvikling.sparesti.dao.ChallengeDAO;
 import idatt2106.systemutvikling.sparesti.dao.ChallengeLogDAO;
+import idatt2106.systemutvikling.sparesti.dao.MilestoneDAO;
 import idatt2106.systemutvikling.sparesti.dto.ChallengeDTO;
+import idatt2106.systemutvikling.sparesti.exceptions.BankConnectionErrorException;
+import idatt2106.systemutvikling.sparesti.exceptions.NotFoundInDatabaseException;
 import idatt2106.systemutvikling.sparesti.mapper.ChallengeMapper;
 import idatt2106.systemutvikling.sparesti.repository.ChallengeLogRepository;
 import idatt2106.systemutvikling.sparesti.repository.ChallengeRepository;
 
 import java.util.logging.Logger;
 
+import idatt2106.systemutvikling.sparesti.repository.MilestoneRepository;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
@@ -22,8 +27,11 @@ import java.util.List;
 public class ChallengeService {
 
   private final Logger logger = Logger.getLogger(ChallengeService.class.getName());
-  private ChallengeRepository challengeRepository;
-  private ChallengeLogRepository challengeLogRepository;
+  private final TransactionService transactionService;
+  private final ChallengeRepository challengeRepository;
+  private final ChallengeLogRepository challengeLogRepository;
+  private final MilestoneService milestoneService;
+  private final MilestoneRepository dbMilestone;
 
   public ChallengeDTO getChallenge(Long challengeId) {
     return ChallengeMapper.toDTO(challengeRepository.findChallengeDAOByChallengeId(challengeId));
@@ -149,12 +157,65 @@ public class ChallengeService {
     return challengeRepository.save(challengeDAO);
   }
 
-  public ChallengeLogDAO completeChallenge(Long challengeId) {
+  @Transactional
+  public void completeChallengeForCurrentUser(long challengeId, long milestoneId) {
+    // Fetch username from security context
+    final String username = CurrentUserService.getCurrentUsername();
+
+    // Fetch data
+    ChallengeDAO challenge = challengeRepository.findByChallengeIdAndUserDAO_Username(challengeId, username);
+    MilestoneDAO milestone = dbMilestone.findMilestoneDAOByMilestoneIdAndUserDAO_Username(milestoneId, username);
+
+    // Verify existence of the requested challenge
+    if (challenge == null)
+      throw new NotFoundInDatabaseException("No challenge found with the requested ID for the user");
+
+    // Verify existence of the requested milestone
+    if (milestone == null)
+      throw new NotFoundInDatabaseException("No milestone found with the requested ID for the user");
+
+
+
+    // Define transfer amount as the difference between the goal and the current sum
+    long transferAmount = challenge.getGoalSum() - challenge.getCurrentSum();
+
+    // Perform savings transaction
+    boolean success = transactionService.createSavingsTransferForCurrentUser(transferAmount);
+
+    // Verify transaction success
+    if (success)
+      throw new BankConnectionErrorException("Failed to transfer funds to savings");
+
+
+    // Transfer achieved currency to milestone
+    MilestoneDAO savedEntry = milestoneService.increaseMilestonesCurrentSum(milestoneId, challenge.getGoalSum());
+
+
+    // Archive challenge
+    archiveActiveChallenge(challenge.getChallengeId());
+  }
+
+  public ChallengeLogDAO archiveActiveChallenge(Long challengeId) {
+    // Fetch active challenge from database
     ChallengeDAO challengeDAO = challengeRepository.findChallengeDAOByChallengeId(challengeId);
+
+    // Verify existence of the active challenge
+    if (challengeDAO == null)
+      return null;
+
+    // Create log entry from the active challenge
     ChallengeLogDAO challengeLogDAO = createChallengeLog(challengeDAO);
+
+    // Set the archived currency equal to the goal of the active challenge
     challengeLogDAO.setChallengeAchievedSum(challengeLogDAO.getGoalSum());
+
+    // Save challenge log entry to database
+    challengeLogDAO = challengeLogRepository.save(challengeLogDAO);
+
+    // Remove active challenge from database
     challengeRepository.delete(challengeDAO);
-    challengeLogRepository.save(challengeLogDAO);
+
+    // Return the archived entry
     return challengeLogDAO;
   }
